@@ -1,22 +1,21 @@
 using System.Diagnostics.CodeAnalysis;
 using EzDomain.EventSourcing.Domain.Model;
 
-namespace EzDomain.EventSourcing.Tests.TestDoubles
+namespace EzDomain.EventSourcing.Tests.TestDoubles;
+
+[ExcludeFromCodeCoverage]
+internal sealed class TestAggregateRootId
+    : IAggregateRootId
 {
-    [ExcludeFromCodeCoverage]
-    internal sealed class TestAggregateRootId
-        : IAggregateRootId
+    private readonly string _value;
+
+    public TestAggregateRootId(string value)
     {
-        private readonly string _value;
+        _value = value;
+    }
 
-        public TestAggregateRootId(string value)
-        {
-            _value = value;
-        }
-
-        public override string ToString()
-        {
-            return _value;
-        }
+    public override string ToString()
+    {
+        return _value;
     }
 }
