@@ -1,15 +1,14 @@
 using System.Diagnostics.CodeAnalysis;
 using EzDomain.EventSourcing.Domain.Model;
 
-namespace EzDomain.EventSourcing.Tests.TestDoubles
+namespace EzDomain.EventSourcing.Tests.TestDoubles;
+
+[ExcludeFromCodeCoverage]
+internal sealed class UnhandledBehaviorExecuted
+    : Event
 {
-    [ExcludeFromCodeCoverage]
-    internal sealed class UnhandledBehaviorExecuted
-        : Event
+    public UnhandledBehaviorExecuted(string aggregateRootId)
+        : base(aggregateRootId)
     {
-        public UnhandledBehaviorExecuted(string aggregateRootId)
-            : base(aggregateRootId)
-        {
-        }
     }
 }
